SpelledOut.jl Documentation

SpelledOut.jl Documentation

Adding SpelledOut.jl

Documentation

Index

Adding SpelledOut.jl

julia> using Pkg

julia> Pkg.add("SpelledOut")
   Updating registry at `~/.julia/registries/General`
  Resolving package versions...
  Installed SpelledOut ─ v1.0.1
Updating `~/work/SpelledOut.jl/SpelledOut.jl/docs/Project.toml`
  [4728c690] ~ SpelledOut v1.0.1 `~/work/SpelledOut.jl/SpelledOut.jl` ⇒ v1.0.1
Updating `~/work/SpelledOut.jl/SpelledOut.jl/docs/Manifest.toml`
  [4728c690] ~ SpelledOut v1.0.1 `~/work/SpelledOut.jl/SpelledOut.jl` ⇒ v1.0.1

Documentation

Main.SpelledOut.SPELLED_OUT — Method

SPELLED_OUT(number; lang = :en, british = false, dict = :modern)

Spells out a number in words, in uppercase, given a specified language. Current languages to choose from are:

:en

Dictionaries supported include :modern, :british, and :european, and default to the former.

If the british boolean is true (which it is not by default), the programme will add "and" where needed.

Main.SpelledOut.Spelled_Out — Method

Spelled_Out(number; lang = :en, british = false, dict = :modern)

Spells out a number in words, in title-case, given a specified language. Current languages to choose from are:

:en

Dictionaries supported include :modern, :british, and :european, and default to the former.

If the british boolean is true (which it is not by default), the programme will add "and" where needed.

Main.SpelledOut.Spelled_out — Method

Spelled_out(number; lang = :en, british = false, dict = :modern)

Spells out a number in words, starting with a capital letter, given a specified language. Current languages to choose from are:

:en

Dictionaries supported include :modern, :british, and :european, and default to the former.

If the british boolean is true (which it is not by default), the programme will add "and" where needed.

Main.SpelledOut.spelled_out — Method

spelled_out(number; lang = :en, british = false, dict = :modern)

Spells out a number in words, in lowercase, given a specified language. Current languages to choose from are:

:en

Dictionaries supported include :modern, :british, and :european, and default to the former.

If the british boolean is true (which it is not by default), the programme will add "and" where needed.

Examples

julia> spelled_out(12, lang = :en)
"twelve"

julia> spelled_out(112, lang = :en)
"one hundred twelve"

julia> spelled_out(112, lang = :en, british = true)
"one hundred and twelve"

julia> spelled_out(112, lang = :en, dict = :european)
"one hundred twelve"

Index

Main.SpelledOut.SPELLED_OUT

Main.SpelledOut.Spelled_Out

Main.SpelledOut.Spelled_out

Main.SpelledOut.spelled_out