Usage

Main.SpelledOut.SPELLED_OUT — Method

SPELLED_OUT(number; lang = :en, british = false, dict = :modern)

A wrapper for spelled_out. Spells out a number in words, in uppercase, given a specified language.

Main.SpelledOut.Spelled_Out — Method

Spelled_Out(number; lang = :en, british = false, dict = :modern)

A wrapper for spelled_out. Spells out a number in words, in title-case, given a specified language.

Main.SpelledOut.Spelled_out — Method

Spelled_out(number; lang = :en, british = false, dict = :modern)

A wrapper for spelled_out. Spells out a number in words, starting with a capital letter, given a specified language.

Main.SpelledOut.spelled_out — Method

spelled_out(number; lang = :en, british = false, dict = :modern)

Spells out a number in words, in lowercase, given a specified language.

Examples

julia> spelled_out(12, lang = :en)
"twelve"

julia> spelled_out(112, lang = :en)
"one hundred twelve"

julia> spelled_out(112, lang = :en, british = true)
"one hundred and twelve"

julia> spelled_out(112, lang = :en, dict = :european)
"one hundred twelve"